Rename input testing module to inputtest
