+++ filebeat/input/container/input_test.go
@@ -24,7 +24,7 @@ import (
 	"path"
 	"testing"
 
-	inputtest "github.com/elastic/beats/v7/filebeat/input/testing"
+	"github.com/elastic/beats/v7/filebeat/input/inputtest"
 	"github.com/elastic/beats/v7/libbeat/common"
 )
