+++ filebeat/input/inputtest/input.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package testing
+package inputtest
 
 import (
 	"testing"
