+++ filebeat/input/stdin/input_test.go
@@ -22,7 +22,7 @@ package stdin
 import (
 	"testing"
 
-	inputtest "github.com/elastic/beats/v7/filebeat/input/testing"
+	"github.com/elastic/beats/v7/filebeat/input/inputtest"
 	"github.com/elastic/beats/v7/libbeat/common"
 )
