+++ x-pack/filebeat/input/gcppubsub/input_test.go
@@ -22,7 +22,7 @@ package gcppubsub
 import (
 	"testing"
 
-	inputtest "github.com/elastic/beats/v7/filebeat/input/testing"
+	"github.com/elastic/beats/v7/filebeat/input/inputtest"
 	"github.com/elastic/beats/v7/libbeat/common"
 )
